# do not pass -threads option to Cbc if only one thread